requirements.txt
aiohttp==3.7.4.post0
alpha-vantage==2.3.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
Django==3.2.1
djangorestframework==3.12.4
idna==2.10
multidict==5.1.0
pytz==2021.1
requests==2.25.1
schedule==1.1.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.4
yarl==1.6.3
